requirements.txt
pyopenssl
ndg-httpsclient
pyasn1
Mastodon.py
python-twitter
pyyaml
